refactor: 디자인 토큰 관련 계층 이동 shared 계층에서도 common-ui 정의를 위해 사용되어야하는 의존성인 이유로 이동함
