+++ src/shared/common-ui/theme/styled.d.ts
